pyproject.toml
[tool.poetry]
name = "django-test"
version = "1.0.0"
description = ""
authors = ["jhlee3124 <lung3124@gmail.com>"]
readme = "README.md"
packages = [{include = "django_test"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.2.0"
pillow = "^10.0.0"
jupyter = "^1.0.0"
pandas = "^2.1.0"
xlwt = "^1.3.0"
requests = "^2.31.0"
django-bootstrap4 = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
